Revert to using separate session hooks
