+++ client/src/lib/user/hooks/usePinnedSessions.ts
@@ -8,7 +8,7 @@ import useCurrentUserState from './useCurrentUserState';
 const usePinnedSessons = () => {
   const userState = useCurrentUserState();
   const setPinnedSessions = useUserState(state => state.setPinnedSessions);
-  const {logSessionMetricEvent} = useLogSessionMetricEvents();
+  const logSessionMetricEvent = useLogSessionMetricEvents();
 
   const pinnedSessions = useMemo(
     () => userState?.pinnedSessions ?? [],
@@ -34,7 +34,7 @@ const usePinnedSessons = () => {
             expires: dayjs(session.startTime).add(1, 'month').toDate(),
           },
         ]);
-        logSessionMetricEvent('Add Sharing Session To Interested');
+        logSessionMetricEvent('Add Sharing Session To Interested', session);
       }
     },
     [pinnedSessions, setPinnedSessions, logSessionMetricEvent],
