+++ client/src/routes/Session/hooks/useLeaveSession.spec.ts
@@ -22,7 +22,7 @@ jest.mock('../state/state', () =>
 );
 
 const mockConditionallyLogLeaveSessionMetricEvent = jest.fn();
-jest.mock('./useLogSessionMetricEvents', () => () => ({
+jest.mock('./useLogInSessionMetricEvents', () => () => ({
   conditionallyLogLeaveSessionMetricEvent:
     mockConditionallyLogLeaveSessionMetricEvent,
 }));
